backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
motor==3.3.2
pymongo==4.6.0

# Authentication and security
bcrypt==4.1.1
PyJWT==2.8.0
passlib==1.7.4

# Firebase integration
firebase-admin==6.2.0

# Twilio integration
twilio==8.10.3

# HTTP client for weather API
aiohttp==3.9.1
requests==2.31.0

# Data validation and serialization
pydantic==2.5.0
pydantic[email]==2.5.0

# Image processing and ML (for crop analysis)
Pillow==10.1.0
numpy==1.25.2
opencv-python==4.8.1.78
tensorflow==2.15.0
scikit-learn==1.3.2

# Environment variables
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.2

# Date/time utilities
python-dateutil==2.8.2

# File handling
aiofiles==23.2.1

# CORS support (included in FastAPI but explicit)
fastapi-cors==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
twilio>=9.0.0
firebase-admin>=6.2.0
bcrypt>=4.0.0
